Moving inflows and outflows into their own files
